+++ pomp/contrib/urllibtools.py
@@ -30,11 +30,8 @@ class UrllibDownloader(BaseDownloader):
         self.timeout = timeout
 
     def get(self, requests):
-        responses = []
         for request in iterator(requests):
-            response = self._fetch(request)
-            responses.append(response)
-        return responses
+            yield self._fetch(request)
 
     def _fetch(self, request):
         try: